requirements.txt
# Core
streamlit>=1.31.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40

# RAG & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# Input Processing
pytesseract>=0.3.10
Pillow>=10.0.0
openai-whisper>=20231117

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
